change audio file to be copyright friendly
