+++ client/src/components/PianoSound.js
@@ -1,0 +1,47 @@
+import { Howl } from 'howler';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import {
+  solid,
+  regular,
+  brands,
+} from '@fortawesome/fontawesome-svg-core/import.macro';
+
+export default function PianoSound() {
+  const sound = new Howl({
+    //TODO: Remove direct reference to localhost in file src
+    src: ['http://localhost:8080/public/Piano.mp3'],
+    html5: true,
+    preload: true,
+    loop: true,
+    volume: 1,
+  });
+
+  return (
+    <section className="flex flex-row justify-around">
+      <div>
+        <h1 className="font-body text-2xl text-center">Piano </h1>
+      </div>
+      <div>
+        <button
+          type="button"
+          name="playSound"
+          className="px-3"
+          onClick={() => {
+            sound.play();
+          }}
+        >
+          <FontAwesomeIcon icon={solid('circle-play')} />
+        </button>
+        <button
+          type="button"
+          name="playSound"
+          onClick={() => {
+            sound.pause();
+          }}
+        >
+          <FontAwesomeIcon icon={solid('circle-pause')} />
+        </button>
+      </div>
+    </section>
+  );
+}
